requirements.txt
# Robot Framework Core
robotframework==7.2.2

# Browser Library – Playwright Integration für Robot Framework
# Wichtig: Nach pip install muss noch "rfbrowser init" ausgeführt werden!
# Das lädt die Playwright Browser-Binaries herunter (~300 MB).
# Entspricht "npx playwright install" aus der Playwright TS Welt.
robotframework-browser==19.4.0

# Parallele Testausführung (Bonus)
robotframework-pabot==2.18.0

# Demo Web-Applikation
flask==3.1.0

# Testdaten-Generierung (für CustomKeywords.py)
faker==33.3.0

# HTTP-Requests (für CustomKeywords.py API-Tests)
requests==2.32.3
